requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
python-dateutil==2.9.0.post0
qrcode==8.2
six==1.17.0
tzdata==2025.3
Werkzeug==3.1.6
PyJWT
gunicorn
python-dotenv
